refactor(exporter): extract HTTP handlers + main wiring helpers (PR-1/5)
Slim package main by separating transport / boot / lifecycle concerns.
First of 5 v2.8.0 refactor PRs aimed at code-health before tag.

Layout changes:
  - NEW handlers.go (139 LoC) — healthHandler / readyHandler /
    configViewHandler extracted verbatim from main.go (signatures
    pinned by main_test.go: bare HandlerFunc for /health, closure
    factories for /ready and /api/v1/config)
  - RENAME api_simulate.go → handler_simulate.go (sibling naming;
    content unchanged, simulateHandler signature preserved)
  - main.go 248 → 197 LoC; main() body 80 → 28 LoC

Helpers extracted from main():
  - loadEnvOverrides() — CONFIG_PATH / CONFIG_DIR / LISTEN_ADDR
  - buildMux(manager, collector) — routing table; testable without
    binding a port
  - buildServer(addr, handler) — http.Server with hardened defaults
    (Gosec G112 ReadHeaderTimeout, 8 KiB MaxHeaderBytes, pinned
    timeouts)
  - runUntilSignal(server, stopCh, manager) — SIGINT/SIGTERM block
    + ordered shutdown (close stopCh → manager.Close → server.Shutdown
    with 15s budget). Order matters per v2.7.0 §8.11.2 trap #12.

Pinned by tests (must not change shape):
  - resolveConfigPath() stays in package main with its globals
  - all three handler signatures unchanged

Verification:
  - go build -buildvcs=false ./... clean
  - go test -buildvcs=false -race -count=1 ./... — all 9 packages pass
    (threshold-exporter / cmd/da-{batchpr,guard,parser} /
     internal/{batchpr,guard,parser,profile} / pkg/config)

Sets up the seam for PR-3 (diffAndReload split) and PR-5
(ConfigManager god-object decompose) — handlers.go is now the
isolation boundary that the post-refactor manager lives behind.
